Stub button for non-mobile platforms
